+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Released on Thursday, May 5 2019.
 
 - Reference latest AngleSharp
 - Returns `null` in `GetStyle` if CSS not configured (#15)
+- Added `pointer-events` and fixed border recombination (#16)
 - Added extension helper `SetStyle` to modify all styles of many elements (#22)
 - Introduced special converter for the `src` declaration (#25)
 - Fixed bug regarding CSS grid serialization (#27)